Switch to a cache model in which caches are stored in the layer.
For each item (i.e. shape-datum combination), the layer asks the shape
for any cache it might like to derive from the supplied data to be
provided to a subsequent update call. The layer then stores this against
that shape and datum and passes it to every update for them.

The waveform implementation of this initially uses a single peak cache.