Don't request non-smooth scaling after all; use small overlap instead
The non-smooth scaling request doesn't work on all browsers (only seems
to have any effect in Firefox). We want behaviour to be consistent
everywhere so we'll just have to do things differently. Introduce a
small overlap between tiles, which at least fixes the problem of having
a visible join when smoothing up to the edges of adjacent tiles.